requirements.txt
asgiref==3.6.0
Django==4.1.5
django-jalali-date==1.0.1
jdatetime==4.1.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
